Fix sanity_ok() for prompt_runner.
The prompt_runner is called as soon as the first pane is displayed. But
the client is not setup yet. Every time prompt_runner calls the
tmxr_runner() function, sanity_ok() is called, and in turn the
check_tmux_version script is called. If the version check fails, tmux
display-message is called. The goal is to display acceptable versions
but the call fails because the client is not yet setup.